simplify file API to simple read/write

+++ src/jsvar.c
@@ -925,7 +925,7 @@ void jsvAppendString(JsVar *var, const char *str) {
 }
 
 // Append the given string to this one - but does not use null-terminated strings. returns false on failure (from out of memory)
-bool jsvAppendStringBuf(JsVar *var, const char *str, int length) {
+bool jsvAppendStringBuf(JsVar *var, const char *str, size_t length) {
   assert(jsvIsString(var));
   JsVar *block = jsvLockAgain(var);
   // Find the block at end of the string...
